# src/modules/users/dtos/update-user.dto.ts
import { ApiPropertyOptional } from '@nestjs/swagger';
import { IsString, IsOptional, IsUrl } from 'class-validator';

export class UpdateUserDto {
  @ApiPropertyOptional({
    description: 'Display name of the user',
    example: 'Jane Smith',
  })
  @IsString()
  @IsOptional()
  name?: string;

  @ApiPropertyOptional({
    description: 'URL of the user avatar image',
    example: 'https://example.com/avatars/jane.png',
  })
  @IsUrl()
  @IsOptional()
  avatarUrl?: string;
}