# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/users/profile": {
      "get": {
        "description": "Returns the authenticated user profile (password excluded).",
        "operationId": "UsersController_getProfile",
        "parameters": [],
        "responses": {
          "200": {
            "description": "User profile retrieved successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "email": "jane.doe@example.com",
                      "name": "Jane Doe",
                      "avatarUrl": "https://example.com/avatars/jane.png",
                      "status": "ACTIVE",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T10:00:00.000Z",
                      "deletedAt": null
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Get current user profile",
        "tags": [
          "users"
        ]
      },
      "patch": {
        "description": "Updates the authenticated user profile. Only provided fields are updated.",
        "operationId": "UsersController_updateProfile",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateUserDto"
              },
              "examples": {
                "updateName": {
                  "summary": "Update name only",
                  "value": {
                    "name": "Jane Smith"
                  }
                },
                "updateAvatar": {
                  "summary": "Update avatar URL only",
                  "value": {
                    "avatarUrl": "https://example.com/avatars/new-avatar.png"
                  }
                },
                "updateBoth": {
                  "summary": "Update name and avatar",
                  "value": {
                    "name": "Jane Smith",
                    "avatarUrl": "https://example.com/avatars/jane-smith.png"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "statusCode": 200,
                    "message": "Operation successfully!",
                    "data": {
                      "id": "a1b2c3d4-e5f6-7890-abcd-ef1234567890",
                      "email": "jane.doe@example.com",
                      "name": "Jane Doe",
                      "avatarUrl": "https://example.com/avatars/jane.png",
                      "status": "ACTIVE",
                      "createdAt": "2025-02-14T10:00:00.000Z",
                      "updatedAt": "2025-02-14T10:00:00.000Z",
                      "deletedAt": null
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "Update current user profile",
        "tags": [
          "users"
        ]
      }
    },
...
  },
  "info": {
    "title": "Starter Kit API Documentation",
    "description": "API documentation for Starter Kit",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "jwt": {
        "scheme": "Bearer",
        "bearerFormat": "Bearer",
        "description": "[just text field] Please enter token in following format: Bearer <JWT>",
        "name": "Authorization",
        "type": "http",
        "in": "Header"
      }
    },
    "schemas": {
...
      "UpdateUserDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Display name of the user",
            "example": "Jane Smith"
          },
          "avatarUrl": {
            "type": "string",
            "description": "URL of the user avatar image",
            "example": "https://example.com/avatars/jane.png"
          }
        }
      },
...
    }
  },
  "security": [
    {
      "jwt": []
    }
  ]
}